backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
apscheduler>=3.10.0
requests>=2.31.0
tushare>=1.4.0
pandas>=2.0.0
pytest>=8.0.0
